Change column type for value in scores table

+++ database/migrations/2024_05_31_131021_create_scores_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
                 ->references('id')
                 ->on('criterias')
                 ->onDelete('cascade');
-            $table->decimal('value', 5, 2);
+            $table->integer('value');
             $table->timestamps();
         });
     }
